Start breaking up the monolithic `api.py` file more
