+++ src/flickr_photos_api/api/__init__.py
@@ -4,28 +4,15 @@ import itertools
 import xml.etree.ElementTree as ET
 
 from flickr_url_parser import ParseResult, parse_flickr_url
-import httpx
 from nitrate.xml import (
     find_optional_text,
     find_required_elem,
     find_required_text,
 )
-from tenacity import (
-    retry,
-    retry_if_exception,
-    RetryError,
-    stop_after_attempt,
-    wait_random_exponential,
-)
 
-from .exceptions import (
-    FlickrApiException,
-    InvalidApiKey,
-    InvalidXmlException,
-    LicenseNotFound,
-    ResourceNotFound,
-)
-from .types import (
+from .base import BaseApi as NewBaseApi
+from ..exceptions import LicenseNotFound
+from ..types import (
     CollectionOfElements,
     CollectionOfPhotos,
     Comment,
@@ -43,7 +30,7 @@ from .types import (
     UserInfo,
     user_info_to_user,
 )
-from .utils import (
+from ..utils import (
     parse_date_posted,
     parse_date_taken,
     parse_location,
@@ -52,124 +39,7 @@ from .utils import (
 )
 
 
-def is_retryable(exc: BaseException) -> bool:
-    """
-    Returns True if this is an exception we can safely retry (i.e. flaky
-    or transient errors that might return a different result),or
-    False otherwise.
-    """
-    if isinstance(exc, httpx.HTTPStatusError) and exc.response.status_code == 500:
-        return True
-
-    if isinstance(exc, httpx.ReadTimeout):
-        return True
-
-    if isinstance(exc, InvalidXmlException):
-        return True
-
-    # Sometimes we get an error from the Flickr API like:
-    #
-    #     <err
-    #       code="201"
-    #       msg="Sorry, the Flickr API service is not currently available."
-    #     />
-    #
-    # but this indicates a flaky connection rather than a genuine failure.
-    if (
-        isinstance(exc, FlickrApiException)
-        and isinstance(exc.args[0], dict)
-        and exc.args[0].get("code") == "201"
-    ):
-        return True
-
-    return False
-
-
-class BaseApi:
-    """
-    This is a thin wrapper for calling the Flickr API.
-
-    It doesn't do much interesting stuff; the goal is just to reduce boilerplate
-    in the rest of the codebase, e.g. have the XML parsing in one place rather
-    than repeated everywhere.
-    """
-
-    def __init__(self, *, api_key: str, user_agent: str) -> None:
-        if not api_key:
-            raise ValueError(
-                "Cannot create a client with an empty string as the API key"
-            )
-
-        self.client = httpx.Client(
-            base_url="https://api.flickr.com/services/rest/",
-            params={"api_key": api_key},
-            headers={"User-Agent": user_agent},
-        )
-
-    def call(self, *, method: str, params: dict[str, str] | None = None) -> ET.Element:
-        try:
-            return self._call_api(method=method, params=params)
-        except RetryError as retry_err:
-            retry_err.reraise()
-
-    @retry(
-        retry=retry_if_exception(is_retryable),
-        stop=stop_after_attempt(5),
-        wait=wait_random_exponential(),
-    )
-    def _call_api(self, *, method: str, params: dict[str, str] | None) -> ET.Element:
-        if params is not None:
-            get_params = {"method": method, **params}
-        else:
-            get_params = {"method": method}
-
-        resp = self.client.get(url="", params=get_params, timeout=15)
-        resp.raise_for_status()
-
-        # Note: the xml.etree.ElementTree is not secure against maliciously
-        # constructed data (see warning in the Python docs [1]), but that's
-        # fine here -- we're only using it for responses from the Flickr API,
-        # which we trust.
-        #
-        # However, on occasion I have seen it return error messages in
-        # JSON rather than XML, which causes this method to fail -- make
-        # sure we log the offending text, and allow it to be retried as
-        # a temporary failure.
-        #
-        # [1]: https://docs.python.org/3/library/xml.etree.elementtree.html
-        try:
-            xml = ET.fromstring(resp.text)
-        except ET.ParseError as err:
-            raise InvalidXmlException(
-                f"Unable to parse response as XML ({resp.text!r}), got error {err}"
-            )
-
-        # If the Flickr API call fails, it will return a block of XML like:
-        #
-        #       <rsp stat="fail">
-        #       	<err
-        #               code="1"
-        #               msg="Photo &quot;1211111111111111&quot; not found (invalid ID)"
-        #           />
-        #       </rsp>
-        #
-        # Different API endpoints have different codes, and so we just throw
-        # and let calling functions decide how to handle it.
-        if xml.attrib["stat"] == "fail":
-            errors = find_required_elem(xml, path=".//err").attrib
-
-            # Although I haven't found any explicit documentation of this,
-            # it seems like a pretty common convention that error code "1"
-            # means "not found".
-            if errors["code"] == "1":
-                raise ResourceNotFound(method, params)
-            elif errors["code"] == "100":
-                raise InvalidApiKey(message=errors["msg"])
-            else:
-                raise FlickrApiException(errors)
-
-        return xml
-
+class BaseApi(NewBaseApi):
     def _to_parsed_element(self, photo_elem: ET.Element) -> ParsedElement:
         """
         Given a <photo> element from a collection response, parse some
